adding authentication4

+++ backend/core/routers/post_routers.py
@@ -1,9 +1,11 @@
-from fastapi import APIRouter,Depends,UploadFile,Form,File
+from fastapi import APIRouter,Depends,UploadFile,Form,File,HTTPException
 from sqlalchemy.orm import Session
 from typing import Text
 from schemas.posts import show_post
 from core.crud.post_crud import create_new_post,update_post_by_id,delete_post_by_id,get_all_user_posts
 from db.session import get_db
+from routers.login_router import get_current_user
+from db.models.users import User
 from db.models.posts import Post
 
 
@@ -20,7 +22,10 @@ async def Create_new_post_router(owner_id:int,db:Session=Depends(get_db),image:U
     db.refresh(new_post)
 
 @router.put("/update-post/{post_id}")
-def update_post_by_id(post_id:int,post_new_data:dict,db:Session=Depends(get_db)):
+def update_post_by_id(post_id:int,post_new_data:dict,db:Session=Depends(get_db),current_user:User=Depends(get_current_user)):
+    if current_user.id !=Post.owner_id:
+         raise HTTPException(status_code=403, detail="Not authorized to perform this action")
+
     Postnew_data=update_post_by_id(post_id,post_new_data,db)
     if Postnew_data:
           return {"message":f"post with ID {post_id} has been updated"}
@@ -30,7 +35,10 @@ def update_post_by_id(post_id:int,post_new_data:dict,db:Session=Depends(get_db))
   
        
 @router.delete("/post/{post_id}")
-def Delete_post_by_id(post_id:int,db:Session=Depends(get_db)):
+def Delete_post_by_id(post_id:int,db:Session=Depends(get_db),current_user:User=Depends(get_current_user)):
+     if current_user.id !=Post.owner_id:
+          raise HTTPException(status_code=403, detail="Not authorized to perform this action")
+
      Delete_post=delete_post_by_id(post_id,db)
      if Delete_post:
           return {"message":f"post with ID {post_id} has been updated"}
